# mbms/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/attendance/clock-in": {
      "post": {
        "operationId": "AttendanceController_clockIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/attendance/clock-out": {
      "post": {
        "operationId": "AttendanceController_clockOut",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/attendance/mark-absent": {
      "post": {
        "operationId": "AttendanceController_markAbsent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkAbsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/attendance/records": {
      "get": {
        "operationId": "AttendanceController_listRecords",
        "parameters": [
          {
            "name": "filter[companyId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[employeeId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[dateFrom]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[dateTo]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/attendance/summary": {
      "get": {
        "operationId": "AttendanceController_summary",
        "parameters": [
          {
            "name": "filter[companyId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[dateFrom]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[dateTo]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/attendance/shifts": {
      "get": {
        "operationId": "AttendanceController_listShifts",
        "parameters": [
          {
            "name": "filter[companyId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "AttendanceController_createShift",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/attendance/employees/{employeeId}/shift": {
      "patch": {
        "operationId": "AttendanceController_assignShift",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignShiftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/ar/credit-notes": {
      "post": {
        "operationId": "AccountsReceivableController_createCreditNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerCreditNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ar/debit-notes": {
      "post": {
        "operationId": "AccountsReceivableController_createDebitNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDebitNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ar/aging": {
      "get": {
        "operationId": "AccountsReceivableController_aging",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ar/outstanding-invoices": {
      "get": {
        "operationId": "AccountsReceivableController_outstandingInvoices",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ar/statement": {
      "get": {
        "operationId": "AccountsReceivableController_statement",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ar/invoices/{id}/send-reminder": {
      "post": {
        "operationId": "AccountsReceivableController_sendReminder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Morise MBMS API",
    "description": "Morise Holdings Limited Business Management System — Phase 1 API. Base path /api/v1. Staff endpoints take a `staff` bearer token from POST /identity/auth/login; Customer Storefront endpoints take a `customer` bearer token from POST /customer-portal/auth/login.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "staff": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "customer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ClockDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          }
        }
      },
      "MarkAbsentDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "employeeId",
          "date"
        ]
      },
      "CreateShiftDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "pattern": "/^([01]\\d|2[0-3]):[0-5]\\d$/"
          },
          "endTime": {
            "type": "string",
            "pattern": "/^([01]\\d|2[0-3]):[0-5]\\d$/"
          },
          "graceMinutes": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "companyId",
          "name",
          "startTime",
          "endTime"
        ]
      },
      "AssignShiftDto": {
        "type": "object",
        "properties": {
          "shiftId": {
            "type": "string"
          }
        }
      },
...
      "CreateCustomerCreditNoteDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "creditDate": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "reason": {
            "type": "string"
          },
          "appliedToInvoiceId": {
            "type": "string"
          },
          "offsetAccountId": {
            "type": "string"
          },
          "financialPeriodId": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "customerId",
          "creditDate",
          "amount",
          "offsetAccountId",
          "financialPeriodId"
        ]
      },
      "CreateCustomerDebitNoteDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "debitDate": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "reason": {
            "type": "string"
          },
          "offsetAccountId": {
            "type": "string"
          },
          "financialPeriodId": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "customerId",
          "debitDate",
          "amount",
          "offsetAccountId",
          "financialPeriodId"
        ]
      },
...
    }
  }
}